% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlitdata-attribute-filter.R
\name{mlitdata_attribute_filter}
\alias{mlitdata_attribute_filter}
\title{Attribute filter}
\usage{
mlitdata_attribute_filter(..., .env = rlang::caller_env())
}
\arguments{
\item{...}{Attribute filter expressions.}

\item{.env}{Environment to evaluate expressions in.}
}
\value{
An \code{mlitdata_query_args} object.
}
\description{
Attribute filter
}
