% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlitdata-location-filter.R
\name{mlitdata_location_filter}
\alias{mlitdata_location_filter}
\alias{mlitdata_location_filter_rectangle}
\alias{mlitdata_location_filter_lat_lon_point}
\alias{mlitdata_location_filter_geodistance}
\title{Location filter}
\usage{
mlitdata_location_filter(bbox = NULL, point = NULL, distance = NULL)

mlitdata_location_filter_rectangle(bbox = NULL)

mlitdata_location_filter_lat_lon_point(point = NULL)

mlitdata_location_filter_geodistance(point = NULL, distance = NULL)
}
\arguments{
\item{bbox}{A \code{bbox} object from sf package.}

\item{point}{A \code{point} object from sf package.}

\item{distance}{A numeric value.}
}
\value{
An \code{mlitdata_query_args} object.
}
\description{
Location filter
}
