% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlitdata-query.R
\name{mlitdata_query}
\alias{mlitdata_query}
\title{Build a query object}
\usage{
mlitdata_query(.name, ..., .args = NULL)
}
\arguments{
\item{.name}{The name of the query.}

\item{...}{The fields of the query.}

\item{.args}{The arguments of the query.}
}
\value{
A \code{mlitdata_query} object.
}
\description{
Build a query object
}
